guozhiju_api_project/requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
asgiref==3.2.3
atomicwrites==1.4.0
attrs==19.3.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2019.9.11
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
concurrent-log-handler==0.9.9
cryptography==3.2.1
ddt==1.2.2
distro==1.5.0
Django==3.0.2
docker==4.4.0
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
elasticsearch==7.9.1
idna==2.8
jsonschema==3.2.0
kafka-python==1.4.6
more-itertools==8.4.0
nb-log==3.4
numpy==1.19.2
paramiko==2.7.2
pika==1.1.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pymongo==3.5.1
PyMySQL==0.9.3
PyNaCl==1.4.0
pyrsistent==0.17.3
pytest==4.5.0
pytest-base-url==1.4.2
python-dotenv==0.15.0
pytz==2019.3
pywin32==227
PyYAML==3.12
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
selenium==3.141.0
six==1.15.0
sqlparse==0.3.0
texttable==1.6.3
tomorrow3==1.1.0
urllib3==1.25.7
wcwidth==0.2.5
websocket-client==0.57.0
xlrd==1.2.0
